Fixed modelView bug, fixed lighting. More in description
Fixed logout button, so that it clears out the password field when logging out. Changed every single widget from On Clicked event to On Pressed through notepad++ solution (genius). Fixed main menu Log out. Changed pause menu and gave it a widget switcher for when the user is online/offline.